Fix missing job error handling. Closes #36.
Calls to /api/job and /job/<job_id> now throw a 404 if
the requested job does not exist. Additionally, if the user
is on a job detail page (/job/<job_id>) when a job is deleted,
they are forced back to the jobs overview page. This prevents
the user from wasting time on a job that has been deleted.