Adjust verify() in Job.py (indentation, logging)
This commit switches a conditional statement to the opposite in order to
reduce the indentation on the following code block, which is
sufficiently large enough to argue neatness.

Also, an error is raised when referencing an invalid JobTask, instead of
a warning with verfify() returning false. This is because in this case,
the data is fundamentally broken.