Rearchitect job expansion to handle graph AND tasks
The original approach did not consider the fact that not only is there a
dag/graph object in a job, but also a list of task -> task object
mappings that is usually synchronized by the job.

This PR adds a snapshot_tasks object, which will be necessary when
making the dag_snapshot as well, this must match all of the inheretid
jobs, just like the dag.

Then, I revised the code to manage both the graph and the task list
independently. The code is pending a py-dag PR, but I'm committing it
for initial review.