Send task snapshot and name delimiter to events
Events (such as emails) need the snapshot version of the run logs in order to
display all the pertinent info you'd want in an email, or in the logs. The
name delimiter also needs to be passed to that the event handlers can reformat
the name to a human readable version if needed.